compilation warnings cleanup
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@1797 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/PluginLibraryLoader.cpp
@@ -136,7 +136,7 @@ int __tulip_select_libs(struct dirent *ent) {
   int idx = strlen(ent->d_name) - suffix_len;
   if (idx < 0) return 0;
   
-  for (int i=0; i < suffix_len; ++i) {
+  for (unsigned long i=0; i < suffix_len; ++i) {
     if ((ent->d_name[idx + i]) != suffix[i]) return 0;
   }
   return 1;
